Class of objects representing plate tectonic models

Meta-object containing paths to a unique plate tectonic model

Usage

# S4 method for class 'platemodel'
initialize(
  .Object,
  rotation = NULL,
  features = NULL,
  name = NULL,
  polygons = NULL
)

Arguments

.Object

Constructor argument (not needed).

rotation

(character) The path to the rotation file.

features

(character) Named vector of features with the paths to the individual files.

name

(character) (Optional) name of the model.

polygons

(character) (Deprecated) The path to the static plate polygon file.

Value

A platemodel class object.

Examples

# path to provided archive
archive <- file.path(
  system.file("extdata", package="rgplates"), 
  "paleomap_v3.zip")
# extract to temporary directory
unzip(archive, exdir=tempdir())
# path to the rotation file
rotPath <- file.path(tempdir(), 
  "PALEOMAP_PlateModel.rot")
# path to the polygons
polPath <- file.path(tempdir(), 
  "PALEOMAP_PlatePolygons.gpml")
# register in R - to be used in reconstruct()
model <- platemodel(rotation=rotPath, features=c("static_polygons"=polPath))