Reconstruct geographic features

Reconstruct the geographic locations from present day coordinates and spatial objects back to their paleo-positions. Each location will be assigned a plate id and moved back in time using the chosen reconstruction model.

Usage

reconstruct(x, ...)

# S4 method for class 'matrix'
reconstruct(
  x,
  age = 0,
  model = "MERDITH2021",
  from = 0,
  listout = TRUE,
  verbose = FALSE,
  enumerate = TRUE,
  chunk = NULL,
  reverse = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = NULL,
  partitioning = "static_polygons",
  check = TRUE,
  warn = TRUE,
  anchor = 0,
  validtime = TRUE
)

# S4 method for class 'data.frame'
reconstruct(x, ...)

# S4 method for class 'numeric'
reconstruct(x, ...)

# S4 method for class 'character'
reconstruct(
  x,
  age,
  model = "MERDITH2021",
  listout = TRUE,
  verbose = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  partitioning = "static_polygons",
  check = TRUE,
  anchor = 0
)

# S4 method for class 'Spatial'
reconstruct(
  x,
  age,
  model,
  listout = TRUE,
  verbose = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = NULL,
  partitioning = "static_polygons",
  check = TRUE,
  validtime = TRUE,
  anchor = 0
)

# S4 method for class 'sf'
reconstruct(
  x,
  age,
  model,
  listout = TRUE,
  verbose = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = NULL,
  gmeta = FALSE,
  partitioning = "static_polygons",
  check = TRUE,
  validtime = TRUE,
  anchor = 0
)

# S4 method for class 'SpatRaster'
reconstruct(
  x,
  age,
  model,
  from = 0,
  listout = TRUE,
  verbose = FALSE,
  plateperiod = NULL,
  check = TRUE,
  validtime = TRUE
)

Arguments

x

The features to be reconstructed. Can be a vector with longitude and latitude representing a single point or a matrix/dataframe with the first column as longitude and second column as latitude. For the online subroutine, the character strings "static_polygons", "coastlines" and "plate_polygons" return static plate polygons, rotated present-day coastlines and topological plates, respectively. For the offline subroutine, it can be a name of the feature set defined in the model object. Some Spatial*, sf and SpatRaster classes are also accepted, although this input is still experimental.

...

arguments passed to class-specific methods.

age

(numeric) is the target age in Ma at which the feature will be reconstructed. Defaults to 0 Ma.

model

(character or platemodel) The reconstruction model. The class of this argument selects the submodule used for reconstruction, a character value will invoke the remote reconstruction submodule and will submit x to the GPlates Web Service. A platemodel class object will call the local-reconstruction submodule. The default is "PALEOMAP". See details for available models.

from

(numeric) The original age of the features to be reconstructed. A single value, defaults to 0Ma. Only used with the online reconstruction module.

listout

(logical)If multiple ages are given, the output can be returned as a list if listout = TRUE.

verbose

(logical) Should call URLs (remote submodule) or console feedback (local-submodule) be printed?

enumerate

(logical) Should be all coordinate/age combinations be enumerated and reconstructed (set to TRUE by default)? FALSE is applicable only if the number of rows in x is equal to the number elementes in age. Then a point will be reconstructed to the age that has the same index in age as the row of the coordinates in x. List output is not available in this case.

chunk

(numeric) Deprected argument of the online reconstruction method. Ignored.

reverse

(logical) Argument of the remote reconstruction submodule. The flag to control the direction of reconstruction. If reverse = TRUE, the function will calculate the present-day coordinates of the given paleo-coordinates, with age setting the target. Not recommended, kept only for compatibility with the GPlates Web Service. Using from instead of age will automatically trigger reverse reconstruction.

path.gplates

(character) Argument of the local reconstruction submodule. In case the GPlates executable file is not found at the coded default location, the full path to the executable (gplates-<ver>.exe on Windows) can be entered here. e.g. "C:/gplates_2.3.0_win64/gplates.exe".

cleanup

(logical) Argument of the local reconstruction submodule. Should the temporary files be deleted immediately after reconstructions?

dir

(character) Argument of the local reconstruction submodule. Directory where the temporary files of the reconstruction are stored (defaults to a temporary directory created by R). Remember to toggle cleanup if you want to see the files.

plateperiod

(logical) Deprecated argument, renamed to validtime for higher compatibility with the GPlates Web Service.

partitioning

(character) Argument of the local reconstruction submodule, which feature collection of the tectonic model should be used to assing plate IDs to the features? It defaults to "static_polygons".

check

(logical) Should the validity of the entries for the GWS checked with the information stored in gws? (default: TRUE)

warn

(character) Argument of the online reconstruction submodule, used in reverse-reconstructions (calculation of present-day coordinates from paleocoordinates). If set to `TRUE` (default), the function will produce a warning when paleocoordinates are not assigned to any of the paritioning polygons (missing values are returned for these). When set to `FALSE`, the warnings will not be displayed.

anchor

(character) The Plate ID of the anchored plate (defaults to 0). This is the 'anchored_plate_id' parameter of the GPlates Web Service.

validtime

(logical) Argument of the local reconstuction submodule. Should the durations of the plates be forced on the partitioned feature? If these are set to TRUE and the plate duration estimates are long, then you might lose some data. This is the inverse of the ignore.valid.time argument of the GWS.

gmeta

(logical) Argument of the local reconstruction submodule, in the case, when sf objects are supplied. Should the metadata produced by GPlates be included in the output object?

Value

A numeric matrix if x is a numeric, matrix or data.frame, or Spatial* class objects, depending on input. NULL in case no model is specified.

Details

The function implements two reconstruction submodules, which are selected with the model argument:

If model is a character entry, then the reconstruct() function uses the GPlates Web Service (https://gwsdoc.gplates.org/, remote reconstruction submodule). The available reconstruction models for this submodule are (as of 2024-02-02):

"CAO2024" (Cao et al., 2024) for static polygons, coastlines, and topologies (0-1800 Ma).

"ALFONSO2024" (Alfonso et al., 2024) for static polygons, coastlines, and topologies (0-170 Ma).

"MULLER2022" (Müller et al., 2022) for coastlines and static plate polygons (0-1000 Ma).

"ZAHIROVIC2022" (Zahirovic et al., 2022) for static polygons, coastlines, and topologies (0-410 Ma).

"MERDITH2021" (Merdith et al., 2021, default) for coastlines and static plate polygons (0-1000 Ma).

"CLENNET2020" (Clennet et al., 2022) for static polygons, coastlines, and topologies (0-170 Ma).

"MULLER2019" (Müller et al., 2019) for coastlines and static plate polygons. (0-250 Ma).

"MULLER2016" (Muller et al., 2016) for coastlines and topological plate polygons (0-230 Ma).

"MATTHEWS2016_mantle_ref" (Matthews et al., 2016) for coastlines and topological plate polygons (0-410 Ma).

"MATTHEWS2016_pmag_ref" (Matthews et al., 2016) for coastlines and topological plate polygons (0-410 Ma).

"SETON2012" (Seton et al., 2012) for coastlines and topological plate polygons (0-200 Ma).

"GOLONKA" (Wright et al. 2013) for coastlines only (0-550 Ma).

"PALEOMAP" (Scotese, 2016) for coastlines only (0-1100 Ma).

"TorsvikCocks2017" (Torsvik and Cocks, 2017) for coastlines (0-540 Ma). Uses a mantle reference frame by default. For climatically sensitive analyses use a paleomagnetic reference frame, which you can toggle by setting the anchor parameter to 1 from the default 0.

"RODINIA2013" (Li et al., 2012) for coastlines (530-1100 Ma).

If model is a platemodel class object, then the function will try to use the GPLates desktop application (https://www.gplates.org/) to reconstruct the coordinates (local reconstruction submodule). Plate models are available in chronosphere with the fetch function. See datasets for the available models. The function will try to find the main GPlates executable in its default installation directory. If this does not succeed, use path.gplates to enter the full path to the GPlates executable as a character string.

References

Alfonso, C. P., Müller, R. D., Mather, B., & Anthony, M. (2025). Spatio-temporal copper prospectivity in the American Cordillera predicted by positive-unlabeled machine learning. Geological Society of America Bulletin, 137(1-2), 702–711.

Cao, X., Collins, A. S., Pisarevsky, S., Flament, N., Li, S., Hasterok, D., & Müller, R. D. (2024). Earth’s tectonic and plate boundary evolution over 1.8 billion years. Geoscience Frontiers, 15(6), 101922.

Clennett, E. J., Sigloch, K., Mihalynuk, M. G., Seton, M., Henderson, M. A., Hosseini, K., Mohammadzaheri, A., Johnston, S. T., & Müller, R. D. (2020). A quantitative tomotectonic plate reconstruction of western North America and the eastern Pacific basin. Geochemistry, Geophysics, Geosystems, 21(8), e2020GC009117.

Matthews, K. J., Maloney, K. T., Zahirovic, S., Williams, S. E., Seton, M., & Müller, R. D. (2016). Global plate boundary evolution and kinematics since the late Paleozoic. Global and Planetary Change, 146, 226–250. https://doi.org/10.1016/j.gloplacha.2016.10.002

Merdith, A. S., Williams, S. E., Collins, A. S., Tetley, M. G., Mulder, J. A., Blades, M. L., Young, A., Armistead, S. E., Cannon, J., Zahirovic, S., & Müller, R. D. (2021). Extending full-plate tectonic models into deep time: Linking the Neoproterozoic and the Phanerozoic. Earth-Science Reviews, 214, 103477. https://doi.org/10.1016/j.earscirev.2020.103477

Li, Z.-X., Evans, D. A., & Halverson, G. P. (2013). Neoproterozoic glaciations in a revised global palaeogeography from the breakup of Rodinia to the assembly of Gondwanaland. Sedimentary Geology, 294, 219–232.

Müller, R. D., Seton, M., Zahirovic, S., Williams, S. E., Matthews, K. J., Wright, N. M., … Cannon, J. (2016). Ocean Basin Evolution and Global-Scale Plate Reorganization Events Since Pangea Breakup. Annual Review of Earth and Planetary Sciences, 44(1), 107–138. https://doi.org/10.1146/annurev-earth-060115-012211

Müller, R. D., Zahirovic, S., Williams, S. E., Cannon, J., Seton, M., Bower, D. J., Tetley, M. G., Heine, C., Le Breton, E., Liu, S., Russell, S. H. J., Yang, T., Leonard, J., and Gurnis, M. (2019), A global plate model including lithospheric deformation along major rifts and orogens since the Triassic. Tectonics, vol. 38, https://doi.org/10.1029/2018TC005462.

Müller, R. D., Flament, N., Cannon, J., Tetley, M. G., Williams, S. E., Cao, X., Bodur, Ö. F., Zahirovic, S., and Merdith, A.: A tectonic-rules-based mantle reference frame since 1 billion years ago – implications for supercontinent cycles and plate–mantle system evolution, Solid Earth, 13, 1127–1159, https://doi.org/10.5194/se-13-1127-2022, 2022.

Scotese, C. R. (2016). PALEOMAP PaleoAtlas for GPlates and the PaleoData Plotter Program. http://www.earthbyte.org/paleomap‐ paleoatlas‐for‐gplates

Seton, M., Müller, R. D., Zahirovic, S., Gaina, C., Torsvik, T., Shephard, G., … Chandler, M. (2012). Global continental and ocean basin reconstructions since 200Ma. Earth-Science Reviews, 113(3–4), 212–270. https://doi.org/10.1016/j.earscirev.2012.03.002

Torsvik and Cocks (2017). Earth History and Palaeogeography. Cambridge University Press, 317 pp.

Wright, N., Zahirovic, S., Müller, R. D., & Seton, M. (2013). Towards community-driven paleogeographic reconstructions: integrating open-access paleogeographic and paleobiology data with plate tectonics. Biogeosciences, 10(3), 1529–1541. https://doi.org/10.5194/bg-10-1529-2013

Zahirovic, S., Eleish, A., Doss, S., Pall, J., Cannon, J., Pistone, M., Tetley, M. G., Young, A., & Fox, P. (2022). Subduction and carbonate platform interactions. Geoscience Data Journal, 9(2), 371–383.

Examples

# With the web service 
# simple matrices
# replace model with desired choice
reconstruct(matrix(c(95, 54), nrow=1), 140, model=NULL)
#> No model was specified.
#> NULL

# points reconstruction
xy <-cbind(long=c(95,142), lat=c(54, -33))
reconstruct(xy, 140, model=NULL)
#> No model was specified.
#> NULL