Calculate velocities of plate tectonic movements

Queries to return meshes of tectonic plate velocities.

Usage

velocities(x, ...)

# S4 method for class 'missing'
velocities(x, ...)

# S4 method for class 'character'
velocities(
  x,
  age,
  model,
  domain = "longLatGrid",
  type = "MagAzim",
  output = "data.frame",
  polecrop = TRUE,
  verbose = FALSE,
  check = TRUE
)

Arguments

x

character: What should the velocities be calculated for? If nothing is given (i.e. signature(x="missing")) the argument defaults to the only currently working feature collection, the "static_polygons". This is expected to be expanded in the future.

...

Arguments of class-specific methods.

age

numeric: The age in millions of years at which the velocities are to be returned.

model

character: The name of the tectonic model. Similar to that of reconstruct.

domain

character: Either "longLatGrid" or "healpix". "longLatGrid" returns the velocites with the domain of a regular, one-by-one degree longitude-latitude grid. "healpix" will return velocities with the domain of an icosahedral, nearly equidistant grid.

type

character: The type of velocity format that is to be returned, either magnitude and azimuth (type="MagAzim") or easting and northing velocity vectors (type="east_north"). Both result in two variables.

output

character: The class name of the output to be returned. Either data.frame or SpatRaster. The latter requires the terra extension (suggested) and is only available with domain="longLatGrid".

polecrop

logical: Only applicable if output="SpatRaster". The original velocity values are provided as a grid-registered raster, which forces the extent of the raster to be beyond the regular [-180, 180] longitude and [-90, 90] domain, producing warnings when the SpatRaster is used. The default cellraster=TRUE resamples this raster to a native, cell-registered grid. This is an issue only with latitudes, so they get cropped by default. Setting this argument to FALSE will skip cropping.

verbose

logical: Are you interested in more messages?

check

logical Should the validity of the entries for the GWS checked with the information stored in gws? (default: TRUE)

Value

Velocities of tectonic movements. If output="data.frame" then the function returns a data.frame with the longitude, latitude, the two velocity variables and the plate ids they belong to. If output="SpatRaster" then the output will be a multilayered SpatRaster object.

Details

The function returns a mesh of velocities: two variables, either magnitude (mm/year) and azimuth (rad): type="MagAzim" or easting and northing velocity vectors (mm/year): type="east_north". Currently only the online method is supported using the GPlates Web Service (internet connection is required). Available models are in the gws object, and can be provided with arguments similar to reconstruct.

Examples

# dummy example,
# set model to the desired model string, e.g. model="MERDITH2021"
velocities("static_polygons", age=45, model=NULL)
#> No model was specified.
#> NULL