Reconstruct geographic features

Reconstruct the geographic locations from present day coordinates and spatial objects back to their paleo-positions. Each location will be assigned a plate id and moved back in time using the chosen reconstruction model.

Usage

reconstruct(x, ...)

# S4 method for matrix
reconstruct(
  x,
  age,
  model = "MERDITH2021",
  listout = TRUE,
  verbose = FALSE,
  enumerate = TRUE,
  chunk = 200,
  reverse = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = TRUE,
  partitioning = "static_polygons"
)

# S4 method for data.frame
reconstruct(x, ...)

# S4 method for numeric
reconstruct(x, ...)

# S4 method for character
reconstruct(
  x,
  age,
  model = "MERDITH2021",
  listout = TRUE,
  verbose = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = FALSE,
  partitioning = "static_polygons"
)

# S4 method for Spatial
reconstruct(
  x,
  age,
  model,
  listout = TRUE,
  verbose = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = FALSE,
  partitioning = "static_polygons"
)

# S4 method for sf
reconstruct(
  x,
  age,
  model,
  listout = TRUE,
  verbose = FALSE,
  path.gplates = NULL,
  cleanup = TRUE,
  dir = NULL,
  plateperiod = FALSE,
  gmeta = FALSE,
  partitioning = "static_polygons"
)

Arguments

x

are the features to be reconstructed. Can be a vector with longitude and latitude representing a single point or a matrix/dataframe with the first column as longitude and second column as latitude, or a SpatialPolygonsDataFrame class object. For the online subroutine, the character strings "static_polygons", "coastlines" and "plate_polygons" return static plate polygons, rotated present-day coastlines and topological plates, respectively. For the offline subroutine, it can be a name of the feature set defined in the model object.

...

arguments passed to class-specific methods.

age

(numeric) is the age in Ma at which the points will be reconstructed

model

(character or platemodel) The reconstruction model. The class of this argument selects the submodule used for reconstruction, a character value will invoke the remote reconstruction submodule and will submit x to the GPlates Web Service. A platemodel class object will call the local-reconstruction submodule. The default is "PALEOMAP". See details for available models.

listout

(logical)If multiple ages are given, the output can be returned as a list if listout = TRUE.

verbose

(logical) Should call URLs (remote submodule) or console feedback (local-submodule) be printed?

enumerate

(logical) Should be all coordinate/age combinations be enumerated and reconstructed (set to TRUE by default)? FALSE is applicable only if the number of rows in x is equal to the number elementes in age. Then a point will be reconstructed to the age that has the same index in age as the row of the coordinates in x. List output is not available in this case.

chunk

(numeric) Argument of the remote reconstruction submodule. Single integer, the number of coordinates that will be queried from the GPlates in a single go.

reverse

(logical) Argument of the remote reconstruction submodule. The flag to control the direction of reconstruction. If reverse = TRUE, the function will calculate the present-day coordinates of the given paleo-coordinates.

path.gplates

(character) Argument of the local reconstruction submodule. In case the GPlates executable file is not found at the coded default location, the full path to the executable (gplates-<ver>.exe on Windows) can be entered here.

cleanup

(logical) Argument of the local reconstruction submodule. Should the temporary files be deleted immediately after reconstructions?

dir

(character) Argument of the local reconstruction submodule. Directory where the temporary files of the reconstruction are stored (defaults to a temporary directory created by R). Remember to toggle cleanup if you want to see the files.

plateperiod

(logical) Argument of the local reconstuction submodule. Should the durations of the plates be forced on the partitioned feature? If these are set to TRUE and the plate duration estimates are long, then you might lose some data.

partitioning

(character) Argument of the local reconstruction submodule, which feature collection of the tectonic model should be used to assing plate IDs to the features? It defaults to "static_polygons".

gmeta

(logical) Argument of the local reconstruction submodule, in the case, when sf objects are supplied. Should the metadata produced by GPlates be included in the output object?

Value

A numeric matrix if x is a numeric, matrix or data.frame, or Spatial* class objects, depending on input. NULL in case no model is specified.

Details

The function implements two reconstruction submodules, which are selected with the model argument:

If model is a character entry, then the reconstruct() function uses the GPlates Web Service (https://gws.gplates.org/, remote reconstruction submodule). The available reconstruction models for this submodule are (as of 2023-06-29):

"SETON2012" (Seton et al., 2012) for coastlines and topological plate polygons (0-200 Ma).

"RODINIA2013" (Li et al., 2012) for coastlines (530-1100 Ma).

"MULLER2016" (Muller et al., 2016) for coastlines and topological plate polygons (0-230 Ma).

"GOLONKA" (Wright et al. 2013) for coastlines only (0-550 Ma).

"PALEOMAP" (Scotese, 2016) for coastlines only (0-1100 Ma).

"MATTHEWS2016_mantle_ref" (Matthews et al., 2016) for coastlines and topological plate polygons (0-410 Ma).

"MATTHEWS2016_pmag_ref" (Matthews et al., 2016) for coastlines and topological plate polygons (0-410 Ma).

"MULLER2019" (Müller et al., 2019) for coastlines and static plate polygons. (0-250 Ma).

"MERDITH2021" (Merdith et al., 2021, default) for coastlines and static plate polygons (0-1000 Ma).

"MULLER2022" (Müller et al., 2022) for coastlines and static plate polygons (0-1000 Ma).

If model is a platemodel class object, then the function will try to use the GPLates desktop application (https://www.gplates.org/) to reconstruct the coordinates (local reconstruction submodule). Plate models are available in chronosphere with the fetch function. See datasets for the available models. The function will try to find the main GPlates executable in its default installation directory. If this does not succeed, use path.gplates to enter the full path to the GPlates executable as a character string.

References

Matthews, K. J., Maloney, K. T., Zahirovic, S., Williams, S. E., Seton, M., & Müller, R. D. (2016). Global plate boundary evolution and kinematics since the late Paleozoic. Global and Planetary Change, 146, 226–250. https://doi.org/10.1016/j.gloplacha.2016.10.002

Andrew S. Merdith, Simon E. Williams, Alan S. Collins, Michael G. Tetley, Jacob A. Mulder, Morgan L. Blades, Alexander Young, Sheree E. Armistead, John Cannon, Sabin Zahirovic, R. Dietmar Müller, (2021). Extending full-plate tectonic models into deep time: Linking the Neoproterozoic and the Phanerozoic, Earth-Science Reviews, Volume 214, 2021, 103477, ISSN 0012-8252, https://doi.org/10.1016/j.earscirev.2020.103477.

Müller, R. D., Seton, M., Zahirovic, S., Williams, S. E., Matthews, K. J., Wright, N. M., … Cannon, J. (2016). Ocean Basin Evolution and Global-Scale Plate Reorganization Events Since Pangea Breakup. Annual Review of Earth and Planetary Sciences, 44(1), 107–138. https://doi.org/10.1146/annurev-earth-060115-012211

Müller, R. D., Zahirovic, S., Williams, S. E., Cannon, J., Seton, M., Bower, D. J., Tetley, M. G., Heine, C., Le Breton, E., Liu, S., Russell, S. H. J., Yang, T., Leonard, J., and Gurnis, M. (2019), A global plate model including lithospheric deformation along major rifts and orogens since the Triassic. Tectonics, vol. 38, https://doi.org/10.1029/2018TC005462.

Müller, R. D., Flament, N., Cannon, J., Tetley, M. G., Williams, S. E., Cao, X., Bodur, Ö. F., Zahirovic, S., and Merdith, A.: A tectonic-rules-based mantle reference frame since 1 billion years ago – implications for supercontinent cycles and plate–mantle system evolution, Solid Earth, 13, 1127–1159, https://doi.org/10.5194/se-13-1127-2022, 2022.

Scotese, C. R. (2016). PALEOMAP PaleoAtlas for GPlates and the PaleoData Plotter Program. http://www.earthbyte.org/paleomap‐ paleoatlas‐for‐gplates

Seton, M., Müller, R. D., Zahirovic, S., Gaina, C., Torsvik, T., Shephard, G., … Chandler, M. (2012). Global continental and ocean basin reconstructions since 200Ma. Earth-Science Reviews, 113(3–4), 212–270. https://doi.org/10.1016/j.earscirev.2012.03.002

Wright, N., Zahirovic, S., Müller, R. D., & Seton, M. (2013). Towards community-driven paleogeographic reconstructions: integrating open-access paleogeographic and paleobiology data with plate tectonics. Biogeosciences, 10(3), 1529–1541. https://doi.org/10.5194/bg-10-1529-2013

Examples

# With the web service 
# simple matrices
# replace model with desired choice
reconstruct(matrix(c(95, 54), nrow=1), 140, model=NULL)
#> No model was specified.
#> NULL

# points reconstruction
xy <-cbind(long=c(95,142), lat=c(54, -33))
reconstruct(xy, 140, model=NULL)
#> No model was specified.
#> NULL